+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add "Hat" spell.
  - Add "lantern" schematic. Add several of the newer spells (Fly, Cluster Bomb, etc) to the enchant list.
  - Add "Simulate 3D" spell.
+ - Allow clicking (or hotbar button) on an empty space to clear active spell. Useful for enchanted swords and other tools.
  
 ## 2.9.8.1